Extract FilePermissions into its own file
